+++ wltp/vmax.py
@@ -155,7 +155,7 @@ def calc_v_max(
         return wots_df
 
     wot[c.n] = wot.index
-    wot[c.p_avail] = wot[c.p] * (1.0 - f_safety_margin)
+    wot[c.p_avail] = wot[c.p] * (1 - f_safety_margin)
 
     ## Scan gears from high --> low-4 but stop at most on 2nd gear.
     #  TODO: apply selective logic for ng-x gears from Heinz-DB?