+++ CHANGES.rst
@@ -95,6 +95,8 @@ v1.0.0.dev6  (3-Aug-2019): PY3.5 only & real work!
   - break: cycle-part limits are plain lists-of-limits (not list-of-pairs).
   - break: flatten model, merging `vehicle` & `params` properties up to root.
   - drop: don't add a sample WOT in base_model if not given.
+  - Inverse safety margin from 0.9 --> 0.1, and stop supporting ngear array 
+    (had been left like that since the phase 1a).
 
 - algo: